requirements.txt
asgiref==3.6.0
colorama==0.4.6
Django==4.2.1
Pillow==9.5.0
pypng==0.20220715.0
qrcode==7.4.2
sqlparse==0.4.4
typing-extensions==4.6.1
tzdata==2023.3
